pyproject.toml
[project]
name = "database-crash-test"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.2.2",
    "google-cloud>=0.34.0",
    "google-cloud-storage>=3.1.0",
    "h5py>=3.13.0",
    "nrel-rex>=0.3.1",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "pyyaml>=6.0.2",
]
